r2466/sofa-dev : FIX: chainHybrid with the RegularGrid. Setting of initial position done using the TransformationVisitor
WANING:  * function get of GNodes will not get objects under desactivated nodes. Like that, we can desactivate a node at initialization without problem due to the topology. Collision models under desactivated node won't be passed to the collision pipeline

	 * "Data" and "DataPtr" can be not displayed in the GUI. It is done by putting in initData, or initDataPtr as last argument a boolean. If it is false, the Data won't be displayed. It can be changed dynamically. ( setDisplayed(bool) ).

FIX: warnings

 //partial sync (15/18 changes)//

+++ modules/sofa/component/collision/DistanceGridCollisionModel.cpp
@@ -775,7 +775,7 @@ DistanceGrid* DistanceGrid::load(const std::string& filename, double scale, int 
         ny = mesh.distmap->ny;
         nz = mesh.distmap->nz;
         ftl::Vec3f fpmin = ftl::transform(mesh.distmap->mat,ftl::Vec3f(0,0,0))*(float)scale;
-        ftl::Vec3f fpmax = ftl::transform(mesh.distmap->mat,ftl::Vec3f(nx-1,ny-1,nz-1))*(float)scale;
+        ftl::Vec3f fpmax = ftl::transform(mesh.distmap->mat,ftl::Vec3f((float)(nx-1),(float)(ny-1),(float)(nz-1)))*(float)scale;
         pmin = Coord(fpmin.ptr());
         pmax = Coord(fpmax.ptr());
         std::cout << "Copying "<<nx<<"x"<<ny<<"x"<<nz<<" distance grid in <"<<pmin<<">-<"<<pmax<<">"<<std::endl;
